{{cookiecutter.project_slug}}/requirements/development.txt
-r base.txt
django-debug-toolbar>=1.4
Werkzeug>=0.11.2
django-autofixture==0.12.1

#Testing
selenium
django-nose
nose
coverage
coveralls
flake8
pep8